+++ clients/src/test/java/org/apache/kafka/clients/consumer/RangeAssignorTest.java
@@ -27,8 +27,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 
-import static org.apache.kafka.clients.consumer.internals.PartitionAssignor.dynamicMember;
-import static org.apache.kafka.clients.consumer.internals.PartitionAssignor.MemberInfo;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
@@ -36,7 +34,6 @@ public class RangeAssignorTest {
 
     private RangeAssignor assignor = new RangeAssignor();
     private String consumerId = "consumer";
-    private MemberInfo memberInfo = dynamicMember(consumerId);
     private String topic = "topic";
 
     @Test
